Test_Shop/backend/src/app/requirements.txt
pymongo~=4.1.1
Flask~=2.1.1
Werkzeug~=2.1.1
numpy~=1.19.5
Flask-API.yandex~=0.6.2.1
